% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_ggb.R
\name{cv_ggb}
\alias{cv_ggb}
\title{Cross Validation to Select Lambda for GGB Estimator}
\usage{
cv_ggb(x, fit, g, errfun = NULL, nfolds = 5, ...)
}
\arguments{
\item{x}{n-by-p data matrix}

\item{fit}{object of class ggb (as outputted by \code{\link{ggb}})}

\item{g}{seed graph (should match that used to compute fit)}

\item{errfun}{a user-specified function measuring the loss incurred by
estimating \code{est} (first argument) when the true covariance matrix
is \code{true} (second argument).  Default: Squared Frobenius norm.}

\item{nfolds}{number of folds for cross-validation (default: 5)}

\item{...}{additional arguments (passed to ggb)}
}
\value{
\describe{
\item{errs: }{A \code{nlam}-by-\code{nfolds} matrix of errors.
\code{errs[l, j]} is error incurred in using \code{lamlist[l]} on fold \code{j}}
\item{m: }{CV error error for each value of lambda.}
\item{se: }{Standard error (estimated over folds) for each value of lambda}
\item{lambda_best: }{Value of \code{lambda} minimizing CV error.}
\item{ibest: }{Index of \code{lambda} minimizing CV error.}
\item{lambda_1se: }{Selected value of lambda using the one-standard-error
rule, a common heuristic that favors a sparser model when there isn't strong
 evidence against it.}
\item{i1se: }{Index of \code{lambda} of one-standard-error rule.}
}
}
\description{
Performs nfold-fold cross validation to select the tuning parameter for GGB.
Takes \code{fit} object created by \code{\link{ggb}} and computes CV error at
the sequence of lambda values from that object.
}
\examples{
set.seed(123)
n <- 20
p <- 10
g <- igraph::graph.lattice(c(5, 2))
x <- matrix(rnorm(n * p), n, p)
S <- stats::cov(x)
fit <- ggb(S, g, type = "local")
cv <- cv_ggb(x, fit, g)
plot(cv)
Sighat <- fit$Sig[[cv$i1se]]
}
\seealso{
\code{\link{ggb}}
}

