% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_graph_banded.R
\name{generate_gb_covariance}
\alias{generate_gb_covariance}
\title{Generate a Graph-Banded Covariance Matrix}
\usage{
generate_gb_covariance(g, b, sigma = 0.01, cor = TRUE)
}
\arguments{
\item{g}{a seed graph (with p vertices)}

\item{b}{a p-vector of local bandwidths}

\item{sigma}{sd of independent error to add to each node}

\item{cor}{whether to scale rows/cols to have variance 1}
}
\description{
Returns a p-by-p matrix covariance matrix that is banded with respect to the
graph \code{g} and with local bandwidths given in the p-vector \code{b}.
}
\examples{
g <- igraph::graph.lattice(c(5, 4))
b <- rep(1, 20)
Sig <- generate_gb_covariance(g, b)
}

