% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb_global_nopsd.R
\name{ggb_global_nopsd}
\alias{ggb_global_nopsd}
\title{Compute the Global GGB estimator without psd constraint}
\usage{
ggb_global_nopsd(S, g, lambda = NULL, w = NULL, flmin = 0.01, nlam = 20)
}
\arguments{
\item{S}{p-by-p sample covariance matrix}

\item{g}{seed graph (not necessarily connected)}

\item{lambda}{positive tuning parameter(s)}

\item{w}{weights from b = 1 to diam(g). Default is sqrt( group size )}

\item{flmin}{ratio of smallest to largest lambda value (ignored if
\code{lambda} is nonnull)}

\item{nlam}{number of lambda values (ignored if \code{lambda} is nonnull)}
}
\description{
Uses the path BCD "closed form" of Yan and Bien (2015) as implemented
in R package \pkg{hsm}.
}

