% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ggb.R
\name{image_covariance}
\alias{image_covariance}
\title{Show the Image of a Covariance Matrix}
\usage{
image_covariance(Sig, sub = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{Sig}{covariance matrix}

\item{sub}{subtitle, default NULL}

\item{xlab}{x label, default NULL}

\item{ylab}{y label, default NULL}

\item{...}{additional arguments to pass to \code{Matrix::image}}
}
\description{
Uses the \code{image} function defined in \pkg{Matrix} package.
}

