% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_ggb.R
\name{make_folds}
\alias{make_folds}
\title{Make folds for cross validation}
\usage{
make_folds(n, nfolds)
}
\arguments{
\item{n}{sample size}

\item{nfolds}{number of folds}
}
\description{
Divides the indices \code{1:n} into \code{nfolds} random folds of about the same size.
}
