% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb_global_nopsd.R
\name{pathprox}
\alias{pathprox}
\title{R Wrapper to C Function}
\usage{
pathprox(y, lam, assign, w = NULL)
}
\arguments{
\item{y}{array of p elements}

\item{lam}{positive tuning parameter}

\item{assign}{array indexing each element in y with node indices (with
0 meaning element is not in the graph)}

\item{w}{array of positive weights, one per node}
}
\value{
b, w, and pen (value of lam * Omega(b) )
}
\description{
Calls C function defined in R package \pkg{hsm}.
}
