% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv.R
\name{plot.cv.ggb}
\alias{plot.cv.ggb}
\title{Plot a GGB-CV object}
\usage{
\method{plot}{cv.ggb}(x, ...)
}
\arguments{
\item{x}{an object of class \code{cv.ggb}, as returned by
\code{\link{cv_ggb}}}

\item{...}{additional arguments (not used)}
}
\description{
This function is adapted from the \pkg{hierNet} R package.
}
