% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ggb.R
\name{plot.ggb}
\alias{plot.ggb}
\title{Plot a GGB object}
\usage{
\method{plot}{ggb}(x, subset = seq_along(x$Sig), ...)
}
\arguments{
\item{x}{ggb object as returned by \code{\link{ggb}}}

\item{subset}{which lambda values to plot, expressed as a subset of indices,
between 1 and \code{length(x$Sig)}}

\item{...}{additional arguments to pass to \code{Matrix::image}}
}
\description{
Uses image function from the \pkg{Matrix} R package.
}
