% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ggb.R
\name{print.ggb}
\alias{print.ggb}
\title{Print output for a GGB object}
\usage{
\method{print}{ggb}(x, ...)
}
\arguments{
\item{x}{ggb object as returned by \code{\link{ggb}}}

\item{...}{additional arguments (not used)}
}
\description{
Print output for a GGB object
}

