% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb_local_nopsd.R
\name{prox_bcd}
\alias{prox_bcd}
\title{R Wrapper to C Function for GGB-Local Penalty Proximal Operator}
\usage{
prox_bcd(R, D, M, lambda = NULL, nlam = 20, flmin = 0.01, maxiter = 100,
  tol = 1e-04, verbose = 1)
}
\arguments{
\item{R}{a symmetric, p by p matrix}

\item{D}{matrix of graph distances}

\item{M}{p-vector of maximal bandwidths}

\item{lambda}{a vector of positive tuning parameters. If NULL, uses
\code{nlam}, \code{flmin} to produce a grid (formation of lambda grid
in this case is done within C)}

\item{nlam}{number of lambda values (ignored if \code{lambda} is nonnull)}

\item{flmin}{ratio of smallest to largest lambda value (ignored if
\code{lambda} is nonnull)}

\item{maxiter}{maximum number of passes through coordinates}

\item{tol}{convergence threshold}

\item{verbose}{level of verbosity in printed output (2, 1, 0)}
}
\description{
Returns sumvv which can be thought of as a (p choose 2)-by-nlam matrix,
 each column being the lower triangle of sum_jb V_jb.
}
\details{
argmin 0.5 * || R - sum_jb V_jb ||_F^2 + lam * sum_jb w_jb || V_jb ||_F
 where V_jb is 0 off of g_jb.
}

