% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggb_local_nopsd.R
\name{ggb_local_nopsd}
\alias{ggb_local_nopsd}
\title{Evaluate the GGB-Local Penalty Proximal Operator Along Grid of Lambda Values}
\usage{
ggb_local_nopsd(S, g, lambda = NULL, w = NULL, flmin = 0.01, nlam = 20,
  max_depths = NULL, maxiter = 500, tol = 1e-04, verbose = 1)
}
\arguments{
\item{S}{a symmetric, p by p matrix}

\item{g}{seed graph}

\item{lambda}{a vector of positive tuning parameters. If NULL, uses
\code{nlam}, \code{flmin} to produce a grid}

\item{w}{weights (currently not implemented)}

\item{flmin}{ratio of smallest to largest lambda value (ignored if
\code{lambda} is nonnull)}

\item{nlam}{number of lambda values (ignored if \code{lambda} is nonnull)}

\item{max_depths}{maximal bandwidths considered.  Should be a p-vector of
non-negative integers. Default is NULL, which is equivalent to taking
max_depths to be >= diameter(g).  (These are referred to as M_j in
the paper.)}

\item{maxiter}{maximum number of passes through coordinates}

\item{tol}{convergence threshold}

\item{verbose}{level of verbosity in printed output (2, 1, 0)}
}
\description{
Returns \code{Sig}, a (length \code{nlam}) list of \code{Matrix} objects,
 with \code{Sig[[l]]} being the solution for the l-th lambda value.
}
